test: separating tests and fixing errors
