+++ src/ansys/geometry/core/designer/design.py
@@ -801,6 +801,7 @@ class Design(Component):
         components: list[Component] | None = None,
         vertices: list[Vertex] | None = None,
         design_curves: list[DesignCurve] | None = None,
+        datum_points: list[DatumPoint] | None = None,
     ) -> NamedSelection:
         """Create a named selection on the active Geometry server instance.
 
@@ -824,6 +825,8 @@ class Design(Component):
             All vertices to include in the named selection.
         design_curves : list[DesignCurve], default: None
             All design curves to include in the named selection.
+        datum_points : list[DatumPoint], default: None
+            All datum points to include in the named selection.
 
         Returns
         -------
@@ -838,12 +841,22 @@ class Design(Component):
         """
         # Verify that at least one entity is provided
         if not any(
-            [bodies, faces, edges, beams, design_points, components, vertices, design_curves]
+            [
+                bodies,
+                faces,
+                edges,
+                beams,
+                design_points,
+                components,
+                vertices,
+                design_curves,
+                datum_points,
+            ]
         ):
             raise ValueError(
                 "At least one of the following must be provided: "
                 "bodies, faces, edges, beams, design_points, components, vertices, "
-                "or design_curves."
+                "design_curves, or datum_points."
             )
 
         named_selection = NamedSelection(
@@ -858,6 +871,7 @@ class Design(Component):
             components=components,
             vertices=vertices,
             design_curves=design_curves,
+            datum_points=datum_points,
         )
 
         self._named_selections[named_selection.name] = named_selection
